:recycle: refactor: update image usage with next static import
