+++ apps/recipe/src/components/Image.tsx
@@ -3,7 +3,16 @@ import NextImage, { ImageProps } from "next/image";
 
 type ImagePropKey = keyof ImageProps | "layout";
 
-const SKIPPED_PROPS: ImagePropKey[] = ["width", "height", "src", "alt", "layout"];
+const SKIPPED_PROPS: ImagePropKey[] = [
+  "width",
+  "height",
+  "src",
+  "alt",
+  "layout",
+  "quality",
+  "placeholder",
+  "priority",
+];
 
 const Image = chakra(NextImage, {
   shouldForwardProp: (prop: ImagePropKey) => SKIPPED_PROPS.includes(prop),
