--- apps/recipe/src/constants/image.ts
@@ -1,6 +1,0 @@
-export const DEFAULT_IMAGE = "/assets/demo.jpg";
-
-export const LOGO = "/favicon/logo.png";
-export const LOGO_TRANSPARENT = "/favicon/logo_transparent.png";
-export const LOGO_HORIZONTAL = "/favicon/logo_horizontal.png";
-export const LOGO_HORIZONTAL_TRANSPARENT = "/favicon/logo_horizontal_transparent.png";
