+++ apps/recipe/src/pages/index.tsx
@@ -1,9 +1,9 @@
 import type { GetStaticPropsResult, NextPage } from "next";
 import React from "react";
 
+import demo from "@/../public/assets/demo.jpg";
 import Intro from "@/components/Intro";
 import Landing from "@/components/Landing";
-import { DEFAULT_IMAGE } from "@/constants/image";
 import { getRecipes } from "@/services/recipe";
 import type { Recipe } from "@/types/recipe";
 
@@ -14,7 +14,7 @@ interface HomeProps {
 /* TODO: refactor Landing*/
 const Home: NextPage<HomeProps> = ({ recipes }) => (
   <>
-    <Landing imageUrl={recipes[0]?.image[0]?.formats.small.url || DEFAULT_IMAGE} />
+    <Landing imageUrl={recipes[0]?.image[0]?.formats.small.url || demo} />
     <Intro recipes={recipes} />
   </>
 );
